Factor iter_senaite_catalogs and rebuild_index into upgrade utils
Both helpers are generic enough to be reused by future upgrade steps,
so move them out of v02_07_000 into senaite.core.upgrade.utils and
import them back from there. No behaviour change.
